+++ Demo/src/main/java/com/braintreepayments/demo/MainFragment.java
@@ -31,7 +31,7 @@ public class MainFragment extends BaseFragment {
     private Button localPaymentsButton;
     private Button preferredPaymentMethodsButton;
     private Button samsungButton;
-    private Button sepaDebitButton;
+    private Button sepaDirectDebitButton;
 
     @Nullable
     @Override
@@ -46,7 +46,7 @@ public class MainFragment extends BaseFragment {
         localPaymentsButton = view.findViewById(R.id.local_payment);
         preferredPaymentMethodsButton = view.findViewById(R.id.preferred_payment_methods);
         samsungButton = view.findViewById(R.id.samsung_pay);
-        sepaDebitButton = view.findViewById(R.id.sepa_debit);
+        sepaDirectDebitButton = view.findViewById(R.id.sepa_debit);
 
         cardsButton.setOnClickListener(this::launchCards);
         payPalButton.setOnClickListener(this::launchPayPal);
@@ -56,7 +56,7 @@ public class MainFragment extends BaseFragment {
         venmoButton.setOnClickListener(this::launchVenmo);
         preferredPaymentMethodsButton.setOnClickListener(this::launchPreferredPaymentMethods);
         samsungButton.setOnClickListener(this::launchSamsungPay);
-        sepaDebitButton.setOnClickListener(this::launchSEPADebit);
+        sepaDirectDebitButton.setOnClickListener(this::launchSEPADirectDebit);
 
         return view;
     }
@@ -133,8 +133,8 @@ public class MainFragment extends BaseFragment {
         Navigation.findNavController(v).navigate(action);
     }
 
-    public void launchSEPADebit(View v) {
-        NavDirections action = MainFragmentDirections.actionMainFragmentToSepaDebitFragment();
+    public void launchSEPADirectDebit(View v) {
+        NavDirections action = MainFragmentDirections.actionMainFragmentToSepaDirectDebitFragment();
         Navigation.findNavController(v).navigate(action);
     }
 }
