+++ Demo/src/main/res/layout/fragment_main.xml
@@ -112,7 +112,7 @@
             android:layout_width="match_parent"
             android:layout_height="wrap_content"
             android:layout_weight="1"
-            android:text="@string/sepa_debit_button"
+            android:text="@string/sepa_direct_debit_button"
             android:textSize="12sp" />
 
         <Button
