+++ Demo/src/main/res/values/strings.xml
@@ -17,7 +17,7 @@
     <string name="ideal_bank_icon">iDEAL bank icon</string>
     <string name="local_payment_button">Local Payment</string>
     <string name="preferred_payment_methods_button">Pref Payment Methods</string>
-    <string name="sepa_debit_button">SEPA Debit</string>
+    <string name="sepa_direct_debit_button">SEPA Direct Debit</string>
     <string name="cards">Credit or Debit Cards</string>
     <string name="samsung_pay">Samsung Pay</string>
     <string name="create_transaction">Create a Transaction</string>
