+++ settings.gradle
@@ -13,4 +13,4 @@ include ':SharedUtils'
 include ':ThreeDSecure'
 include ':TestUtils'
 include ':Demo'
-include ':SEPADebit'
+include ':SEPADirectDebit'
